deployments/models/MiniCPM-Llama3-V-2_5-vllm/requirements.txt
triton
fastapi
starlette
vllm
ray
flash_attn
pillow
pyngrok